Change to BigDecimal

+++ src/main/java/classes/model/Item.java
@@ -2,9 +2,11 @@ package classes.model;
 
 import jdk.nashorn.internal.ir.annotations.Immutable;
 
+import java.math.BigDecimal;
+
 @Immutable
 public class Item {
     public String name;
     public int quantity;
-    public Double price;
+    public BigDecimal price;
 }
